// src/pages/Tokenomics/types.ts
export interface Tokenomics {
  totalSupply: string;
  circulatingSupply: string;
  burned: string;
  marketCap: string;
  fullyDilutedMarketCap: string;
  holders: string;
  transferTax: string;
  liquidityTax: string;
  marketingTax: string;
  devTax: string;
  liquidityPair: string;
  liquidityLocked: string;
  contractAddress: string;
  launchDate: string;
  initialPrice: string;
  allTimeHigh: string;
  allTimeLow: string;
}